Adds support for random ambience messages
This patch set adds support for a table of randomly selected ambience
messages. The strings are stored in the Arduino FLASH (program) memory
to allow more and/or longer strings.

Moved most of the debug message strings into program memory also to
conserve data space.

Also added support for using the A6 and A7 pins on the Arduino UNO,
MICRO, and NANO and other ATMEGA328 based boards as either "called" or
"off_hook" inputs, reading them with analogRead(). These two pins will
require an external pull-up resistor (for active low) or a pull-down
(for active-high).

Digital pin 13 is problematic to use as an input on many Arduino boards
due to an on-board active-high LED circuit. The internal pull-up on the
digital pin just can't pull this pin HIGH. To use D13 as an input will
require an external pull-up resistor.